AG-9212 - Fix theme examples.
