+++ packages/ag-charts-website/src/content/docs/themes/_examples/advanced-theme-cartesian/index.html
@@ -1,0 +1,3 @@
+<div class="wrapper">
+    <div id="myChart"></div>
+</div>
