+++ packages/ag-charts-website/src/content/docs/themes/_examples/advanced-theme-cartesian/main.ts
@@ -1,4 +1,5 @@
 import { AgChart, AgChartOptions, AgChartTheme } from "ag-charts-community"
+import { getData } from "./data"
 
 var myTheme: AgChartTheme = {
   baseTheme: "ag-default-dark",
@@ -11,24 +12,9 @@ var myTheme: AgChartTheme = {
       title: {
         fontSize: 24,
       },
-    },
-    cartesian: {
       padding: {
         left: 70,
         right: 70,
-      },
-      series: {
-        line: {
-          marker: {
-            shape: "circle",
-          },
-        },
-        bar: {
-          label: {
-            enabled: true,
-            color: "white",
-          },
-        },
       },
       axes: {
         category: {
@@ -49,7 +35,22 @@ var myTheme: AgChartTheme = {
         },
       },
     },
-    polar: {
+    line: {
+      series: {
+        marker: {
+          shape: "circle",
+        },
+      },
+    },
+    bar: {
+      series: {
+        label: {
+          enabled: true,
+          color: "white",
+        },
+      },
+    },
+    pie: {
       padding: {
         top: 40,
         bottom: 40,
@@ -58,35 +59,25 @@ var myTheme: AgChartTheme = {
         position: "left",
       },
       series: {
-        pie: {
-          calloutLabel: {
-            enabled: true,
-          },
-          calloutLine: {
-            colors: ["gray"],
-          },
+        calloutLabel: {
+          enabled: true,
+        },
+        calloutLine: {
+          colors: ["gray"],
         },
       },
     },
   },
 }
 
-var data = [
-  { label: "Android", v1: 5.67, v2: 8.63, v3: 8.14, v4: 6.45, v5: 1.37 },
-  { label: "iOS", v1: 7.01, v2: 8.04, v3: 2.93, v4: 6.78, v5: 5.45 },
-  { label: "BlackBerry", v1: 7.54, v2: 1.98, v3: 9.88, v4: 1.38, v5: 4.44 },
-  { label: "Symbian", v1: 9.27, v2: 4.21, v3: 2.53, v4: 6.31, v5: 4.44 },
-  { label: "Windows", v1: 2.8, v2: 1.908, v3: 7.48, v4: 5.29, v5: 8.8 },
-]
-
-var chartOptions1: AgChartOptions = {
+var options: AgChartOptions = {
   theme: myTheme,
-  container: document.getElementById("chart1"),
+  container: document.getElementById("myChart"),
   autoSize: true,
   title: {
     text: "Cartesian Chart Theming",
   },
-  data: data,
+  data: getData(),
   series: [
     {
       type: "bar",
@@ -118,22 +109,4 @@ var chartOptions1: AgChartOptions = {
   ],
 }
 
-var chartOptions2: AgChartOptions = {
-  theme: myTheme,
-  container: document.getElementById("chart2"),
-  autoSize: true,
-  title: {
-    text: "Polar Chart Theming",
-  },
-  data: data,
-  series: [
-    {
-      type: "pie",
-      angleKey: "v4",
-      calloutLabelKey: "label",
-    },
-  ],
-}
-
-var chart1 = AgChart.create(chartOptions1)
-var chart2 = AgChart.create(chartOptions2)
+AgChart.create(options)
