+++ packages/ag-charts-website/src/content/docs/themes/_examples/advanced-theme-cartesian/styles.css
@@ -4,10 +4,6 @@
     flex-direction: column;
 }
 
-#chart1 {
-    height: 50%;
-}
-
-#chart2 {
-    height: 50%;
+#myChart {
+    height: 100%;
 }
