--- packages/ag-charts-website/src/content/docs/themes/_examples/advanced-theme/angular/app/app.module.ts
@@ -1,18 +1,0 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { AgChartsAngularModule } from 'ag-charts-angular';
-import { AppComponent } from './app.component';
-
-
-        @NgModule({
-          imports: [
-            BrowserModule,
-            AgChartsAngularModule
-          ],
-          declarations: [
-            AppComponent
-          ],
-          bootstrap: [ AppComponent ]
-        })
-        export class AppModule { }
-
