--- packages/ag-charts-website/src/content/docs/themes/_examples/advanced-theme/angular/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "ag-chart-packages",
-  "description": "NOTE: This package.json file is solely used by Plunker to look up type definitions.",
-  "dependencies": {
-    "@angular/core": "*",
-    "ag-charts-community": "9.0.0-beta.20230911"
-  }
-}
