--- packages/ag-charts-website/src/content/docs/themes/_examples/advanced-theme/reactFunctional/styles.css
@@ -1,17 +1,0 @@
-.wrapper {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-}
-
-#charts {
-    height: 100%;
-}
-
-#chart1 {
-    height: 50%;
-}
-
-#chart2 {
-    height: 50%;
-}
