--- packages/ag-charts-website/src/content/docs/themes/_examples/advanced-theme/typescript/index.html
@@ -1,4 +1,0 @@
-<div class="wrapper">
-    <div id="chart1"></div>
-    <div id="chart2"></div>
-</div>
