--- packages/ag-charts-website/src/content/docs/themes/_examples/advanced-theme/typescript/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "ag-chart-packages",
-  "description": "NOTE: This package.json file is solely used by Plunker to look up type definitions.",
-  "dependencies": {
-    "ag-charts-community": "9.0.0-beta.20230911"
-  }
-}
