--- packages/ag-charts-website/src/content/docs/themes/_examples/advanced-theme/vue3/styles.css
@@ -1,13 +1,0 @@
-.wrapper {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-}
-
-#chart1 {
-    height: 50%;
-}
-
-#chart2 {
-    height: 50%;
-}
