+++ packages/ag-charts-website/src/content/docs/themes/_examples/stock-themes/data.ts
@@ -1,0 +1,10 @@
+export function getData() {
+    return [
+        { label: 'Android', value: 56.9, other: 7 },
+        { label: 'iOS', value: 22.5, other: 8 },
+        { label: 'BlackBerry', value: 6.8, other: 9 },
+        { label: 'Symbian', value: 8.5, other: 10 },
+        { label: 'Bada', value: 2.6, other: 11 },
+        { label: 'Windows', value: 1.9, other: 12 },
+    ];
+}
