+++ packages/ag-charts-website/src/content/docs/themes/_examples/stock-themes/main.ts
@@ -1,24 +1,13 @@
 import { AgChartOptions, AgChartTheme, AgChart } from 'ag-charts-community'
+import { getData } from './data';
 
 const options: AgChartOptions = {
   container: document.getElementById('myChart'),
   theme: 'ag-default-dark',
-  autoSize: true,
-  padding: {
-    left: 70,
-    right: 70,
-  },
   title: {
     text: 'Chart Theme Example',
   },
-  data: [
-    { label: 'Android', value: 56.9, other: 7 },
-    { label: 'iOS', value: 22.5, other: 8 },
-    { label: 'BlackBerry', value: 6.8, other: 9 },
-    { label: 'Symbian', value: 8.5, other: 10 },
-    { label: 'Bada', value: 2.6, other: 11 },
-    { label: 'Windows', value: 1.9, other: 12 },
-  ],
+  data: getData(),
   series: [
     {
       type: 'pie',
@@ -30,7 +19,7 @@ const options: AgChartOptions = {
 
 var chart = AgChart.create(options)
 
-function applyTheme(theme: string | AgChartTheme) {
+function applyTheme(theme: AgChartTheme) {
   options.theme = theme
   AgChart.update(chart, options)
 }
